+++ travel_planner/src/main/java/com/spring/travel_planner/dao/ReviewDAO.java
@@ -1,28 +1,31 @@
 package com.spring.travel_planner.dao;
 
-import java.util.*;
-
-import org.mybatis.spring.SqlSessionTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Repository;
+import java.util.List;
+import java.util.Map;
 
 import com.spring.travel_planner.dto.ReviewDTO;
 
-@Repository
-public class ReviewDAO {
-
-	@Autowired
-	private SqlSessionTemplate sqlSession;
+public interface ReviewDAO {
 	
-	// 리뷰 개수
-	public int mainReviewCnt() {
-		System.out.println("리뷰DAO - reviewCnt");
-		return sqlSession.selectOne("com.spring.travel_planner.dao.ReviewDAO.mainReviewCnt");
-	}
+	// 후기 목록
+	public List<ReviewDTO> reviewList(Map<String, Object> map);
 	
-	// 조회순으로 리뷰게시글 4개만
-	public List<ReviewDTO> mainReviewList(Map<String, Object> map) {
-		System.out.println("리뷰DAO - reviewList");
-		return sqlSession.selectList("com.spring.travel_planner.dao.ReviewDAO.mainReviewList", map);
-	}
+	// 후기 총 건수
+	public int reviewCnt();
+	
+//	// 조회수 증가
+//	public int reviewViews(int r_num);
+//	
+//	// 후기 상세페이지
+//	public String reviewDetail(int num);
+//	
+//	// 후기 수정처리
+//	public int reviewUpdate(int r_num);
+//	
+//	// 후기 삭제처리
+//	public int reviewDelete(int num);
+//	
+//	// 후기 작성
+//	public int reviewInsert(ReviewDTO dto);
+	
 }
