+++ travel_planner/src/main/resources/mappers/reviewMapper.xml
@@ -5,18 +5,14 @@
 
 <mapper namespace="com.spring.travel_planner.dao.ReviewDAO">
 
-<!-- 리뷰 개수 -->
-<select id="mainReviewCnt" resultType="int">
-	SELECT COUNT(*) FROM travle_review_tb
-</select>
 
-<!-- 조회순으로 리뷰게시글 4개만 -->
+<!-- main - 조회순으로 리뷰게시글 4개만 -->
 <select id="mainReviewList" parameterType="java.util.Map" resultType="com.spring.travel_planner.dto.ReviewDTO">
 	<![CDATA[
 	SELECT *
 	  FROM (SELECT R.*
 						  , rownum AS rn
-					FROM (SELECT * FROM travle_review_tb
+					FROM (SELECT * FROM travle_review_tbl
 							  WHERE r_show = 'Y'
 							  ORDER BY r_readCnt DESC) R
 				)
@@ -24,4 +20,22 @@
 	  	 AND rn <= #{end}
 	]]>
 </select>
+
+<!-- 후기 목록 -->
+<select id="reviewList" parameterType="java.util.Map" resultType="com.spring.travel_planner.dto.ReviewDTO">
+	<![CDATA[
+		SELECT *
+	  	  FROM
+	        (SELECT A.*,
+	                rownum AS rn
+	         FROM
+	                (SELECT * FROM travle_review_tbl
+	                  WHERE r_show = 'Y'
+	                  ORDER BY r_num DESC) A
+	        )
+	 	 WHERE rn >= #{start}
+	 	   AND rn <= #{end}
+	]]>
+</select>
+
 </mapper>
