--- travel_planner/src/main/resources/mappers/reviewMapper1.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-	PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-	"http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-	
-<mapper namespace="com.spring.travel_planner.dao.ReviewDAO1">
-
-<!-- 후기 목록 -->
-<select id="reviewList" parameterType="java.util.Map" resultType="com.spring.travel_planner.dto.ReviewDTO1">
-	<![CDATA[
-		SELECT *
-	  	  FROM
-	        (SELECT A.*,
-	                rownum AS rn
-	         FROM
-	                (SELECT * FROM travle_review_tbl
-	                  WHERE r_show = 'Y'
-	                  ORDER BY r_num DESC) A
-	        )
-	 	 WHERE rn >= #{start}
-	 	   AND rn <= #{end}
-	]]>
-</select>
-
-</mapper>
